fix(summaries): publish season as Int64, not Float64
season is a join key across the whole ecosystem and was being published as a
float -- 2023.0 -- in four released tables: team_summaries, passing, rushing and
receiving. All four route through _prepare_for_write, which did
`pl.lit(float(yr))`.

This is a parity port of R's prepare_for_write and R numerics are doubles, so
the float was a faithful translation of the source; it just is not a correct
dtype to PUBLISH. Every other producer site here (io.py, reshape.py,
reshapers.py, summaries_input.py) already emits Int64 -- this was the lone
outlier.

Found from the consuming side: sdv-py's returns-schema drift test compares the
declared loader schema against the real published parquet, and reported
season Int32 -> Float64 across exactly these four datasets after the
2026-07-28 republish.

Parity is unaffected: _parity_helpers.assert_frame_parity casts numeric columns
to Float64 before comparing, so it is dtype-insensitive, and no test pins
season's type. 41 tests pass.

FOLLOW-UP: the published assets stay Float64 until these five tags are rebuilt.
sdv-py's tools/codegen/schemas/loader_schemas.yaml currently declares Float64 to
match the shipped data, and must flip back to Int64 in the same change as the
republish, or its drift gate will go red.

+++ python/cfb_data_build/team_summaries.py
@@ -448,7 +448,12 @@ def _prepare_for_write(df: pl.DataFrame, yr: int, schools: pl.DataFrame) -> pl.D
     df = _clean_rank_columns(df)
     # R computes fbs_class AFTER the select-rename, so reference the renamed cols.
     out = (
-        df.with_columns(season=pl.lit(float(yr)))
+        # season is a JOIN KEY across the ecosystem and must be an integer. This
+        # emitted Float64 (mirroring R's double numerics), publishing season as
+        # 2023.0 in team_summaries AND the passing/rushing/receiving player tables,
+        # which all route through here. Every other producer site (io.py,
+        # reshape.py, reshapers.py, summaries_input.py) already uses Int64.
+        df.with_columns(season=pl.lit(int(yr), dtype=pl.Int64))
         .join(schools, on="pos_team_id", how="left")
         .rename(
             {
